api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client==1.6.9
ollama==0.1.7
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Database drivers (uncomment based on your lab database)
# psycopg2-binary==2.9.7  # PostgreSQL
# mysql-connector-python==8.1.0  # MySQL
# pyodbc==4.0.39  # SQL Server
# sqlite3 is included in Python standard library